Add clock rate sensitivity analysis
The clock rate is an informed prior (made necessary by the lack of temporal signal in the outbreak dataset) that was derived from a sequence dataset that has serially sampled sequences. Because this is an informed prior, we wanted to see how sensitive our results were to our choice of prior. This commit includes the skyline TSV files under different clock models, as well as the plotting script that plots the skyline plot for the three runs that use different clock priors.