Add xmls and log files from clock sensitivity analysis
This commit pushes up the XML files that BEAST would read in as well as the BEAST output log files for the BEAST runs performed as part of the sensitivity analysis we performed to investigate clock rate prior choice. 

Note that reading in the logfiles to Tracer (program in the BEAST suite) will allow you to generate the skyline plots, and export the TSV files, which is how the TSV files that are also in this repo were generated.