Improve descriptions in README
I realized that it was previously a bit hard to understand the directory structure, namely that things are divided up by which virus we were looking at, and then which analyses were performed. This structure was chosen because the types of analyses performed varied between the 3 viruses. I edited the README to make it easier to understand this structure, and better point to where relevant data and results would be.