Add background context to README
I thought it would be better to have a bit more background information about why we did the study, and what questions we were trying to answer, to contextualize the analyses that we performed. This commit adds more of this context, and improves the readability of some of the brief analysis descriptions.